change to bitmap_xor and for_each_set_bit

+++ sensehat.h
@@ -29,7 +29,7 @@ struct sensehat {
 	struct sensehat_joystick {
 		struct platform_device *pdev;
 		struct input_dev *keys_dev;
-		unsigned int prev_states;
+		unsigned long prev_states;
 	} joystick;
 
 	struct sensehat_display {
